perf(hashmap): avoid first-use key length allocation
