+++ pkg/common/hashmap/inthashmap_test.go
@@ -32,7 +32,7 @@ func TestLegacyIteratorFootprintExcludesTransactionalPlan(t *testing.T) {
 		unsafe.Sizeof((*IntHashMap)(nil))+6*sliceSize,
 		unsafe.Sizeof(intHashMapIterator{}))
 	require.Equal(t,
-		3*unsafe.Sizeof((*StrHashMap)(nil))+7*sliceSize,
+		3*unsafe.Sizeof((*StrHashMap)(nil))+6*sliceSize,
 		unsafe.Sizeof(strHashmapIterator{}))
 }
